+++ src/lib.rs
@@ -4,27 +4,28 @@
 //!
 //! ## Examples
 //! ```
+//! use tempfile;
 //! use cacache;
-//! # use tempfile;
-//! # let tmp = tempfile::tempdir().unwrap();
+//! use async_std::task;
+//!
+//! let tmp = tempfile::tempdir().unwrap();
 //! let dir = tmp.path().to_owned();
-//! cacache::put::data(&dir, "key", b"my-data").unwrap();
-//! let data = cacache::get::read(&dir, "key").unwrap();
-//! assert_eq!(data, b"my-data");
+//! task::block_on(async {
+//!   cacache::put::data(&dir, "key", b"my-async-data").await.unwrap();
+//!   let data = cacache::get::data(&dir, "key").await.unwrap();
+//!   assert_eq!(data, b"my-async-data");
+//! })
 //! ```
 //!
-//! You can also use the equivalent async APIs using async/await!
+//! There are also sync APIs available if you don't want to use async/await:
 //! ```
-//! # use tempfile;
 //! use cacache;
-//! # use async_std::task;
+//! # use tempfile;
 //! # let tmp = tempfile::tempdir().unwrap();
 //! let dir = tmp.path().to_owned();
-//! # task::block_on(async {
-//! cacache::async_put::data(&dir, "key", b"my-async-data").await.unwrap();
-//! let data = cacache::async_get::read(&dir, "key").await.unwrap();
-//! assert_eq!(data, b"my-async-data");
-//! # })
+//! cacache::put::data_sync(&dir, "key", b"my-data").unwrap();
+//! let data = cacache::get::data_sync(&dir, "key").unwrap();
+//! assert_eq!(data, b"my-data");
 //! ```
 #![warn(missing_docs, missing_doc_code_examples)]
 
@@ -39,10 +40,6 @@ pub mod get;
 pub mod ls;
 pub mod put;
 pub mod rm;
-
-pub mod async_get;
-pub mod async_put;
-pub mod async_rm;
 
 pub use errors::Error;
 pub use index::Entry;
